Fixes an issue with URL encoding getting undone by remote servers
The default x-www-url-encoded form post would make it so the whole URL was sent as an encoded string. Then when the remote server decoded it, it'd decode everything (the : and // needed to be), but if you had params that needed to be maintained, like a plus in AWS signature, it'd not encode that, and then AWS (and likely other servers) would treat that as a space.

By sending this as multipart, you resolve that in the simpliest way.